정예반 firebase-회원가입&로그인 과제

+++ src/pages/Detail.jsx
@@ -1,30 +1,49 @@
 import React from "react";
+import { useParams, useNavigate } from "react-router-dom";
 import Header from "../common/Header";
 import Container from "../common/Container";
-import { useParams } from "react-router-dom";
-import { useNavigate } from "react-router-dom";
 import { useDispatch, useSelector } from "react-redux";
 import { deletePost } from "../redux/modules/posts";
+import { RenderIfLoggedIn } from "../redux/modules/renderUtils";
+import { checkUserAuthorization } from "../redux/modules/authUtils";
 
 export default function Detail() {
   const posts = useSelector((state) => state.posts);
+  const isLoggedIn = useSelector((state) => state.user.isLoggedIn);
+  const user = useSelector((state) => state.user.user);
   const dispatch = useDispatch();
   const navigate = useNavigate();
 
   const { id } = useParams();
 
+  // 게시물을 아이디로 조회합니다.
   const post = posts.find((post) => post.id === id);
 
+  // 게시물이 존재하지 않는 경우 처리
+  if (!post) {
+    return <div>존재하지 않는 게시물입니다.</div>;
+  }
+
+  // 해당 게시물의 수정 페이지로 이동 (수정 버튼 클릭 시 호출)
   const handleEditClick = () => {
-    navigate(`/edit/${post.id}`);
+    if (isLoggedIn) {
+      if (checkUserAuthorization(user, post.author, navigate)) {
+        navigate(`/edit/${post.id}`);
+      }
+    }
   };
 
+  // 해당 게시물 삭제 (삭제 버튼 클릭 시 호출)
   const handleDelete = () => {
-    const confirmDelete = window.confirm("정말로 삭제하시겠습니까?");
-    if (confirmDelete) {
-      dispatch(deletePost(post.id));
-      navigate("/");
-      alert("삭제되었습니다.");
+    if (isLoggedIn) {
+      if (checkUserAuthorization(user, post.author, navigate)) {
+        const confirmDelete = window.confirm("정말로 삭제하시겠습니까?");
+        if (confirmDelete) {
+          dispatch(deletePost(post.id));
+          navigate("/");
+          alert("삭제되었습니다.");
+        }
+      }
     }
   };
 
@@ -58,33 +77,37 @@ export default function Detail() {
             justifyContent: "end",
           }}
         >
-          <button
-            onClick={handleEditClick}
-            style={{
-              border: "none",
-              padding: "8px",
-              borderRadius: "6px",
-              backgroundColor: "orange",
-              color: "white",
-              cursor: "pointer",
-              marginRight: "6px",
-            }}
-          >
-            수정
-          </button>
-          <button
-            onClick={handleDelete}
-            style={{
-              border: "none",
-              padding: "8px",
-              borderRadius: "6px",
-              backgroundColor: "red",
-              color: "white",
-              cursor: "pointer",
-            }}
-          >
-            삭제
-          </button>
+          <RenderIfLoggedIn isLoggedIn={isLoggedIn}>
+            <button
+              onClick={handleEditClick}
+              style={{
+                border: "none",
+                padding: "8px",
+                borderRadius: "6px",
+                backgroundColor: "orange",
+                color: "white",
+                cursor: "pointer",
+                marginRight: "6px",
+              }}
+            >
+              수정
+            </button>
+          </RenderIfLoggedIn>
+          <RenderIfLoggedIn isLoggedIn={isLoggedIn}>
+            <button
+              onClick={handleDelete}
+              style={{
+                border: "none",
+                padding: "8px",
+                borderRadius: "6px",
+                backgroundColor: "red",
+                color: "white",
+                cursor: "pointer",
+              }}
+            >
+              삭제
+            </button>
+          </RenderIfLoggedIn>
         </div>
       </Container>
     </>
